requirements.txt
pydrive
watchdog
pystray
pillow
pyinstaller